Feat: IdentifableData 분리
